Add separate build tag for pmem

+++ wal/repair_pmem.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build !pmem
+// +build pmem
 
 package wal
 
@@ -22,6 +22,7 @@ import (
 	"path/filepath"
 
 	"go.etcd.io/etcd/pkg/fileutil"
+	"go.etcd.io/etcd/pkg/pmemutil"
 	"go.etcd.io/etcd/wal/walpb"
 
 	"go.uber.org/zap"
@@ -44,9 +45,19 @@ func Repair(lg *zap.Logger, dirpath string) bool {
 
 	rec := &walpb.Record{}
 
-	var decoder *decoder
+	// Check if the last file opened is in pmem
+	pmemaware, err := pmemutil.IsPmemTrue(dirpath)
+	if err != nil {
+		return false
+	}
 
-	decoder = newDecoder(f)
+	var decoder *decoder
+	if pmemaware {
+		pr := pmemutil.OpenForRead(f.Name())
+		decoder = newDecoder(pr)
+	} else {
+		decoder = newDecoder(f)
+	}
 
 	for {
 		lastOffset := decoder.lastOffset()
@@ -73,16 +84,21 @@ func Repair(lg *zap.Logger, dirpath string) bool {
 			return true
 
 		case io.ErrUnexpectedEOF:
-			bf, bferr := os.Create(f.Name() + ".broken")
-			if bferr != nil {
-				if lg != nil {
-					lg.Warn("failed to create backup file", zap.String("path", f.Name()+".broken"), zap.Error(bferr))
-				} else {
-					plog.Errorf("could not repair %v, failed to create backup file", f.Name())
+			var bf io.Writer
+			if pmemaware {
+				// Do nothing here
+			} else {
+				bf, bferr := os.Create(f.Name() + ".broken")
+				if bferr != nil {
+					if lg != nil {
+						lg.Warn("failed to create backup file", zap.String("path", f.Name()+".broken"), zap.Error(bferr))
+					} else {
+						plog.Errorf("could not repair %v, failed to create backup file", f.Name())
+					}
+					return false
 				}
-				return false
+				defer bf.Close()
 			}
-			defer bf.Close()
 
 			if _, err = f.Seek(0, io.SeekStart); err != nil {
 				if lg != nil {
@@ -93,31 +109,52 @@ func Repair(lg *zap.Logger, dirpath string) bool {
 				return false
 			}
 
-			if _, err = io.Copy(bf, f); err != nil {
-				if lg != nil {
-					lg.Warn("failed to copy", zap.String("from", f.Name()+".broken"), zap.String("to", f.Name()), zap.Error(err))
-				} else {
-					plog.Errorf("could not repair %v, failed to copy file", f.Name())
+			if pmemaware {
+				// TODO Throw error when copy is failing on pmem instead of abrupt termination
+				pmemutil.Copy(f.Name(), f.Name()+".broken")
+				/*if _, err = pmemutil.Copy(f.Name(), f.Name()+".broken"); err != nil {
+					if lg != nil {
+						lg.Warn("failed to copy in pmem", zap.String("from", f.Name()+".broken"), zap.String("to", f.Name()), zap.Error(err))
+					} else {
+						plog.Errorf("could not repair %v, failed to copy file in pmem", f.Name())
+					}
+					return false
+				}*/
+				if err = pmemutil.Resize(f.Name(), lastOffset); err != nil {
+					if lg != nil {
+						lg.Warn("failed to truncate pmem file", zap.String("path", f.Name()), zap.Error(err))
+					} else {
+						plog.Errorf("could not repair %v, failed to truncate pmem file", f.Name())
+					}
+					return false
 				}
-				return false
-			}
+			} else {
+				if _, err = io.Copy(bf, f); err != nil {
+					if lg != nil {
+						lg.Warn("failed to copy", zap.String("from", f.Name()+".broken"), zap.String("to", f.Name()), zap.Error(err))
+					} else {
+						plog.Errorf("could not repair %v, failed to copy file", f.Name())
+					}
+					return false
+				}
 
-			if err = f.Truncate(lastOffset); err != nil {
-				if lg != nil {
-					lg.Warn("failed to truncate", zap.String("path", f.Name()), zap.Error(err))
-				} else {
-					plog.Errorf("could not repair %v, failed to truncate file", f.Name())
+				if err = f.Truncate(lastOffset); err != nil {
+					if lg != nil {
+						lg.Warn("failed to truncate", zap.String("path", f.Name()), zap.Error(err))
+					} else {
+						plog.Errorf("could not repair %v, failed to truncate file", f.Name())
+					}
+					return false
 				}
-				return false
-			}
 
-			if err = fileutil.Fsync(f.File); err != nil {
-				if lg != nil {
-					lg.Warn("failed to fsync", zap.String("path", f.Name()), zap.Error(err))
-				} else {
-					plog.Errorf("could not repair %v, failed to sync file", f.Name())
+				if err = fileutil.Fsync(f.File); err != nil {
+					if lg != nil {
+						lg.Warn("failed to fsync", zap.String("path", f.Name()), zap.Error(err))
+					} else {
+						plog.Errorf("could not repair %v, failed to sync file", f.Name())
+					}
+					return false
 				}
-				return false
 			}
 
 			if lg != nil {
